# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Card Scanner Pro API
servers:
  - url: /api/v1
    description: Canonical versioned base path.
  - url: /api
    description: >-
      Legacy base path, retained for backward compatibility and deprecated in favor of
      /api/v1. Responses on this prefix include RFC 8594 Deprecation and Link
      (rel="successor-version") headers.
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: eventId
          in: query
          schema:
            type: ["integer", "null"]
        - name: contactId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead (including stage)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /leads/{id}/activities:
    get:
      operationId: listLeadActivities
      tags: [leads]
      summary: List activities for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivityList"
    post:
      operationId: createLeadActivity
      tags: [leads]
      summary: Log an activity on a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
  /activities/{id}:
    patch:
      operationId: updateLeadActivity
      tags: [leads]
      summary: Update a logged activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadActivityUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
    delete:
      operationId: deleteLeadActivity
      tags: [leads]
      summary: Delete a logged activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leads/{id}/notes:
    get:
      operationId: listLeadNotes
      tags: [leads]
      summary: List notes for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNoteList"
    post:
      operationId: createLeadNote
      tags: [leads]
      summary: Add a note to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNote"
  /notes/{id}:
    patch:
      operationId: updateLeadNote
      tags: [leads]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNote"
    delete:
      operationId: deleteLeadNote
      tags: [leads]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notes/{id}/history:
    get:
      operationId: listLeadNoteHistory
      tags: [leads]
      summary: List a note's append-only edit history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNoteHistoryList"
  /notes/{id}/comments:
    get:
      operationId: listLeadNoteComments
      tags: [leads]
      summary: List threaded comments on a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNoteCommentList"
    post:
      operationId: createLeadNoteComment
      tags: [leads]
      summary: Add a comment to a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteCommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNoteComment"
  /note-comments/{id}:
    patch:
      operationId: updateLeadNoteComment
      tags: [leads]
      summary: Edit a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteCommentUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNoteComment"
    delete:
      operationId: deleteLeadNoteComment
      tags: [leads]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /leads/{id}/tags:
    get:
      operationId: listLeadTags
      tags: [leads]
      summary: List tags on a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      operationId: attachLeadTag
      tags: [leads]
      summary: Attach a tag to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachLeadTagInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /leads/{id}/tags/{tagId}:
    delete:
      operationId: detachLeadTag
      tags: [leads]
      summary: Detach a tag from a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /leads/{id}/communications:
    get:
      operationId: listLeadCommunications
      tags: [leads]
      summary: List logged communications for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunicationList"
    post:
      operationId: logLeadCommunication
      tags: [leads]
      summary: Log a communication (email/phone/whatsapp/calendar) for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
  /leads/{id}/timeline:
    get:
      operationId: getLeadTimeline
      tags: [leads]
      summary: Unified activity timeline for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineList"
  /leads/{id}/assign:
    post:
      operationId: assignLead
      tags: [leads]
      summary: Assign a lead to an owner and/or team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignLeadInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/bulk-assign:
    post:
      operationId: bulkAssignLeads
      tags: [leads]
      summary: Assign many leads at once using a chosen strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAssignResult"
  /leads/{id}/recommend-assignee:
    post:
      operationId: recommendLeadAssignee
      tags: [leads]
      summary: Preview an AI-recommended owner for a lead (does not mutate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendAssigneeInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssigneeRecommendation"
  /leads/{id}/auto-assign:
    post:
      operationId: autoAssignLead
      tags: [leads]
      summary: Auto-assign a lead to the least-loaded member of its team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/{id}/custom-fields:
    get:
      operationId: getLeadCustomFields
      tags: [custom-fields]
      summary: Get custom field values for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFieldValueList"
    put:
      operationId: setLeadCustomFields
      tags: [custom-fields]
      summary: Set custom field values for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCustomFieldValuesInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFieldValueList"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Lead:
      type: object
      required: [id, companyId, stage, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactCompany:
          type: ["string", "null"]
        stage:
          type: string
          enum: [prospect, qualified, proposal_sent, negotiation, won, lost, new, contacted, meeting_scheduled, archived]
        source:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        closingDate:
          type: ["string", "null"]
        probability:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, "null"]
        notes:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        eventName:
          type: ["string", "null"]
        stageId:
          type: ["integer", "null"]
        stageName:
          type: ["string", "null"]
        stageKey:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        createdAt:
          type: string
          format: date-time
        history:
          type: array
          items:
            $ref: "#/components/schemas/LeadHistory"
    LeadList:
      type: object
      required: [leads, total]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
    LeadInput:
      type: object
      properties:
        contactId:
          type: ["integer", "null"]
        stage:
          type: string
          enum: [prospect, qualified, proposal_sent, negotiation, won, lost, new, contacted, meeting_scheduled, archived]
          default: prospect
        source:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        closingDate:
          type: ["string", "null"]
        probability:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        eventId:
          type: ["integer", "null"]
        stageId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
    LeadUpdate:
      type: object
      properties:
        stage:
          type: string
          enum: [prospect, qualified, proposal_sent, negotiation, won, lost, new, contacted, meeting_scheduled, archived]
        source:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        closingDate:
          type: ["string", "null"]
        probability:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        eventId:
          type: ["integer", "null"]
        stageId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
...
    TagList:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
...
    AttachLeadTagInput:
      type: object
      required: [tagId]
      properties:
        tagId:
          type: integer
    LeadActivity:
      type: object
      required: [id, companyId, type, source, occurredAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        leadId:
          type: ["integer", "null"]
        contactId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        type:
          type: string
        source:
          type: string
          enum: [manual, system]
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        metadata: {}
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    LeadActivityList:
      type: object
      required: [activities]
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/LeadActivity"
    LeadActivityInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [call, email, meeting, message, other, note]
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        occurredAt:
          type: ["string", "null"]
          format: date-time
    LeadActivityUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [call, email, meeting, message, other, note]
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        occurredAt:
          type: ["string", "null"]
          format: date-time
    LeadNote:
      type: object
      required: [id, companyId, leadId, body, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        leadId:
          type: integer
        contactId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        body:
          type: string
        mentions:
          type: array
          items:
            type: integer
        isPinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    LeadNoteList:
      type: object
      required: [notes]
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/LeadNote"
    LeadNoteInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
        isPinned:
          type: boolean
          default: false
    LeadNoteUpdate:
      type: object
      properties:
        body:
          type: string
        isPinned:
          type: boolean
...
    LeadNoteHistoryList:
      type: object
      required: [history]
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/LeadNoteHistory"
    LeadNoteComment:
      type: object
      required: [id, companyId, noteId, body, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        noteId:
          type: integer
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        body:
          type: string
        mentions:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    LeadNoteCommentList:
      type: object
      required: [comments]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/LeadNoteComment"
    LeadNoteCommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
    LeadNoteCommentUpdate:
      type: object
      required: [body]
      properties:
        body:
          type: string
...
    AssignLeadInput:
      type: object
      properties:
        assignedToId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        strategy:
          type: string
          enum: [manual, round_robin, load_balanced, availability, territory, ai]
          description: Assignment rule to apply. Defaults to manual.
    BulkAssignInput:
      type: object
      required: [leadIds]
      properties:
        leadIds:
          type: array
          items:
            type: integer
        strategy:
          type: string
          enum: [manual, round_robin, load_balanced, availability, territory, ai]
        assignedToId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
    BulkAssignResult:
      type: object
      required: [results, assigned, failed]
      properties:
        assigned:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            required: [leadId, success, assignedToId, error]
            properties:
              leadId:
                type: integer
              success:
                type: boolean
              assignedToId:
                type: ["integer", "null"]
              error:
                type: ["string", "null"]
...
    TimelineList:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEntry"
...
    CommunicationInput:
      type: object
      required: [channel]
      properties:
        channel:
          type: string
          enum: [email, phone, whatsapp, calendar]
          description: Communication channel. Maps to an activity type (email→email, phone→call, whatsapp→message, calendar→meeting).
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        occurredAt:
          type: ["string", "null"]
          format: date-time
    CommunicationList:
      type: object
      required: [communications]
      properties:
        communications:
          type: array
          items:
            $ref: "#/components/schemas/LeadActivity"
...